Refactor the ShowName/ShowShortName configuration so that it's done on the parent component first (where the validations are made) and set to the child component.
