+++ COMETwebapp/COMETwebapp.csproj
@@ -30,4 +30,10 @@
       <ProjectReference Include="..\COMET.Web.Common\COMET.Web.Common.csproj" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Content Update="wwwroot\BookInputConfiguration.json">
+        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+      </Content>
+    </ItemGroup>
+
 </Project>
