+++ COMETwebapp/wwwroot/BookInputConfiguration.json
